# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "FyAPI",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
...
        "/public/user/id/{ticketID}": {
            "get": {
                "description": "Get public user by ticket ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get public by ticket ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ticket ID",
                        "name": "ticketID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully get public user by ticket id",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse-dto_UserGet"
                        }
                    },
                    "500": {
                        "description": "Failed to get public user by ticket id",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse-any"
                        }
                    }
                }
            }
        },
        "/public/user/{username}": {
            "get": {
                "description": "Get public user by username",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get public by username",
                "parameters": [
                    {
                        "type": "string",
                        "description": "username",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully get public user by username",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse-dto_UserGet"
                        }
                    },
                    "500": {
                        "description": "Failed to get public user by username",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse-any"
                        }
                    }
                }
            }
        },
...
        "/user/me": {
            "get": {
                "description": "Get current logged-in user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get profile",
                "responses": {
                    "200": {
                        "description": "Successfully get current user",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse-dto_UserGet"
                        }
                    },
                    "500": {
                        "description": "Failed to get current user",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse-any"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update current logged-in user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Update user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "username",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "avatar",
                        "name": "avatar",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully update current user",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse-any"
                        }
                    },
                    "400": {
                        "description": "Username already exists",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse-any"
                        }
                    },
                    "500": {
                        "description": "Failed to update current user",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse-any"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.APIResponse-any": {
            "type": "object",
            "properties": {
                "data": {},
                "meta": {
                    "$ref": "#/definitions/dto.ApiMeta"
                }
            }
        },
...
        "dto.APIResponse-dto_UserGet": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.UserGet"
                },
                "meta": {
                    "$ref": "#/definitions/dto.ApiMeta"
                }
            }
        },
        "dto.ApiMeta": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "dto.UserGet": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "picture": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}